example of using RootState
